correcao de detalhes do exercicio 01

+++ 01/ordenacaoComProcessos.c
@@ -15,7 +15,7 @@ int main(int argc, char **argv){
 	int numero1[argc-1];
 	int numero2[argc-1];
   clock_t c1, c2, c3, c4; /* variáveis que contam ciclos do processador */
-  float tmpPai, tmpFilho;
+  long double tmpPai, tmpFilho;
 
 
 	/* verificar quantidade de parametros */
@@ -50,7 +50,7 @@ int main(int argc, char **argv){
       c2 = clock();
 
       tmpFilho = (float)(c2 - c1)*1000/CLOCKS_PER_SEC; // tempo de execução em milissegundos
-      printf("\nTempo gasto na ordenacao quick sort: %f\n", tmpFilho);
+      printf("\nTempo gasto na ordenacao quick sort: %Lf\n", tmpFilho);
       exit(0);
     }
     else {/*processo pai*/
@@ -71,7 +71,7 @@ int main(int argc, char **argv){
       c4 = clock();
 
       tmpPai = (float)(c4 - c3)*1000/CLOCKS_PER_SEC; // tempo de execução em milissegundos
-      printf("\nTempo gasto na ordenacao simples: %f\n\n", tmpPai);
+      printf("\nTempo gasto na ordenacao simples: %Lf\n\n", tmpPai);
 
       exit(0);
     }
@@ -110,15 +110,3 @@ void quickSort(int valor[], int esquerda, int direita){
     quickSort(valor, i, direita);
   }
 }
-
-
-
-
-
-
-
-// int i;
-//   for (i = 0; i < 10; i++)
-//   {
-//     printf("%d ", numeros[i]);
-//   }
